Добавляет Модуляция
